#156 - got rid of unnecessary fault-type abstraction
