--- angular-client/src/pages/fault-page/shared/fault-type/bms-fault-type.model.ts
@@ -1,11 +1,0 @@
-import { Fault } from '../indiv-fault/fault.model';
-import { FaultType, FaultTypeModel } from './fault-type.model';
-
-export class BMSFaultType implements FaultTypeModel {
-  subscribeToFaults(): Fault<BMSFaultType> {
-    // TODO
-    throw new Error('Method not implemented.');
-  }
-  name = FaultType.BMS;
-  recordedFaults: Fault<BMSFaultType>[] = [];
-}
