--- angular-client/src/pages/fault-page/shared/fault-type/charger-fault-type.model.ts
@@ -1,11 +1,0 @@
-import { Fault } from '../indiv-fault/fault.model';
-import { FaultType, FaultTypeModel } from './fault-type.model';
-
-export class ChargerFaultType implements FaultTypeModel {
-  subscribeToFaults(): Fault<ChargerFaultType> {
-    // TODO
-    throw new Error('Method not implemented.');
-  }
-  name = FaultType.Charger;
-  recordedFaults: Fault<ChargerFaultType>[] = [];
-}
