--- angular-client/src/pages/fault-page/shared/fault-type/mpu-fault-type.model.ts
@@ -1,11 +1,0 @@
-import { Fault } from '../indiv-fault/fault.model';
-import { FaultType, FaultTypeModel } from './fault-type.model';
-
-export class MPUFaultType implements FaultTypeModel {
-  subscribeToFaults(): Fault<MPUFaultType> {
-    // TODO
-    throw new Error('Method not implemented.');
-  }
-  name = FaultType.MPU;
-  recordedFaults: Fault<MPUFaultType>[] = [];
-}
